# Python OnFrameworkData performance and slice.Get()