--- a/Tests/QuantConnect.Tests.csproj
+++ b/Tests/QuantConnect.Tests.csproj
@@ -659,6 +659,7 @@
     <Compile Include="Python\MethodOverloadTests.cs" />
     <Compile Include="Python\PortfolioCustomModelTests.cs" />
     <Compile Include="Python\PythonPackagesTests.cs" />
+    <Compile Include="Python\PythonSliceGetByTypeTest.cs" />
     <Compile Include="Python\PythonThreadingTests.cs" />
     <Compile Include="Python\PythonWrapperTests.cs" />
     <Compile Include="Python\SecurityCustomModelTests.cs" />