<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- AlgorithmPythonWrapper will directly call base OnFrameworkData()
  implementation skipping going through python and it's overhead
- Small performance improvement for adding Tick data points into a Ticks
  collection
- For python always wrap `Slice` with `PythonSlice`, so that `slice.Get()`
  works even when no custom data is present, adding test.

Performance tests running python `RegressionAlgorithm`
`PR`
104.52 seconds at 162k data points per second. Processing total of 16,892,474 data points.
101.16 seconds at 167k data points per second. Processing total of 16,892,474 data points.
103.73 seconds at 163k data points per second. Processing total of 16,892,474 data points.
`Master`
139.46 seconds at 121k data points per second. Processing total of 16,892,474 data points.
135.86 seconds at 124k data points per second. Processing total of 16,892,474 data points.
136.09 seconds at 124k data points per second. Processing total of 16,892,474 data points.

Performance Sampling MASTER
![image](https://user-images.githubusercontent.com/18473240/80416067-e5887f80-88a9-11ea-9a9c-c6eddb76a03e.png)
Performance Sampling PR
![image](https://user-images.githubusercontent.com/18473240/80425648-749d9380-88ba-11ea-92d8-760cde77159d.png)


#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Related to https://github.com/QuantConnect/Lean/issues/3737
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Faster is always better
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit, regression and performance tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->